chore: Update dependencies and remove unused code
